+++ app/variation/__init__.py
@@ -1,0 +1,4 @@
+from flask import Blueprint
+variation = Blueprint('variation', __name__)
+
+from . import views
