+++ app/variation/actions.py
@@ -1,0 +1,3 @@
+from ..utils import get_db_data, calculate_table, \
+    get_cmd_by_gene, get_cmd_by_regin, login_require
+from ..db_const import get_head_cmd
